Matrix: convert dual test to gtest
