+++ src/lib/matrix/test/CMakeLists.txt
@@ -23,7 +23,6 @@ set(tests
     hatvee
     least_squares
     upperRightTriangle
-    dual
     pseudoInverse
 )
 
@@ -43,5 +42,6 @@ endforeach()
 px4_add_unit_gtest(SRC MatrixAssignmentTest.cpp)
 px4_add_unit_gtest(SRC MatrixAttitudeTest.cpp)
 px4_add_unit_gtest(SRC MatrixCopyToTest.cpp)
+px4_add_unit_gtest(SRC MatrixDualTest.cpp)
 px4_add_unit_gtest(SRC MatrixSparseVectorTest.cpp)
 px4_add_unit_gtest(SRC MatrixUnwrapTest.cpp)
